pyproject.toml
[project]
name = "product-viz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "playwright>=1.58.0",
    "plotly>=6.5.2",
    "requests>=2.32.5",
]
